@@ -1,7 +1,10 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using Discord;
-using ConfigReadingLib;
+using BearMLLib;
+using BearMLLib.Serialization.Conversion;
 
 namespace MusicBear.Core
 {
@@ -11,58 +14,37 @@ namespace MusicBear.Core
         public static char Prefix { get; }
         public static string Game { get; set; }
         public static bool SonginStatus { get; }
-        public static string HelpText { get; }
         public static UserStatus Status { get; }
+        public static string HelpText { get; }
 
         static Config()
         {
             try
             {
-                // load help message
-                if(File.Exists("Help.txt"))
-                    HelpText = File.ReadAllText("Help.txt");
-                else
-                    HelpText = "Help message does not exist";
+                var provider = new ConversionProvider(typeof(UserStatus),
+                    str => str.ToLower() switch
+                    {
+                        "online" => UserStatus.Online,
+                        "idle" => UserStatus.Idle,
+                        "donotdisturb" => UserStatus.DoNotDisturb,
+                        "invisible" => UserStatus.Invisible,
+                        _ => throw new Exception("Incorrect status format.")
+                    },
+                    null);
 
-                // load config
-                var reader = new ConfigReader("AppConfig.conf");
-
-                // token
-                Token = reader.GetValue("token");
-
-                // prefix
-                var prefix = reader.GetValue("prefix");
-                if (prefix.Length == 1)
-                    Prefix = Convert.ToChar(prefix);
-                else
-                    throw new Exception("Prefix setting in \"AppConfig.conf\" should be only one character");
-
-                // game
-                Game = reader.GetValue("game");
-
-                // songinstatus
-                var songinstatus = reader.GetValue("songinstatus");
-                if (string.Compare(songinstatus, "True", ignoreCase: true) == 0)
-                    SonginStatus = true;
-                else if (string.Compare(songinstatus, "False", ignoreCase: true) == 0)
-                    SonginStatus = false;
-                else
-                    throw new Exception("SonginStatus setting in \"AppCofig.xml\" isn't correct");
-
-                // status
-                var status = reader.GetValue("status");
-                Status = status.ToLower() switch
-                {
-                    "online" => UserStatus.Online,
-                    "idle" => UserStatus.Idle,
-                    "donotdisturb" => UserStatus.DoNotDisturb,
-                    "invisible" => UserStatus.Invisible,
-                    _ => throw new Exception("Status setting in \"AppCofig.xml\" isn't correct")
-                };
+                var reader = new BearML("config.txt", new[] { provider });
+                
+                Token = reader.GetContent<string>("token");
+                Prefix = reader.GetContent<char>("prefix");
+                Game = reader.GetContent<string>("game");
+                SonginStatus = reader.GetContent<bool>("songinstatus");
+                Status = reader.GetContent<UserStatus>("status");
+                HelpText = reader.GetContent<string>("help");
+                if (string.IsNullOrEmpty(HelpText)) HelpText = "Help message is empty.";
             }
             catch (Exception ex)
             {
-                MessageHandler.Service.AddMessage("Message", ex.Message);
+                MessageHandler.Service.Add("Message", ex.Message);
                 AppControl.Exit();
             }
         }
