@@ -1,12 +1,7 @@
 ﻿using System;
-using System.Threading.Tasks;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 using MusicBear.Core;
 
-namespace MusicBear
-{
-    public class Program
-    {
-        public static async Task Main(string[] args) => 
-            await new EntryPoint().EntryAsync();
-    }
-}
+await new EntryPoint().EntryAsync();
